Updated the License
Added some additional text to counter a bit of the helplessness narrative, and evoke the verbal narrative of being ok about it. Somebody else's norm is not a standard and the best standard you can set are the ones that worked for you. I hope that adding a bit of self-reflection to it helps to establish the fact one can do something without being totally reliant.